Share common client tests in AbstractClientTest
Default, apache, and okhttp client tests extend the Abstract test,
so that we can avoid duplicated code while fully testing each
implementation.
